# refactor: Use OpenDAL API to avoid extra seek call