I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

This PR will refactor the read metadata implement by adopting OpenDAL's native API instead of seek on reader.

Benchmark:

Test upon a 100MiB parquet file hosted on minio.

```rust
old/old                 time:   [4.5618 ms 4.6360 ms 4.7137 ms]
new/new                 time:   [1.4413 ms 1.5201 ms 1.6057 ms]
```

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/12524)
<!-- Reviewable:end -->