@Xuanwo 2023-08-21
![image](https://github.com/datafuselabs/databend/assets/5351546/6b90de93-e2fc-472a-b438-4f062f2ae798)

Tpch looks good, but hits seems OOM, let's try again.