@Xuanwo 2023-08-21
Q0 is 186x slower, seems related to cache changes, let's try again.