@dantengsky 2023-08-21
![Untitled](https://github.com/datafuselabs/databend/assets/22081156/8f627ede-f4f3-4941-8992-7ecc721f3601)

regression of q19 looks weird, this query should benefit from the improvement of this PR