@Xuanwo 2023-08-22
I conducted a test in the development environment and found that this PR does not have a significant impact on real use cases. I recommend merging this PR initially to determine if it resolves the issue of unexpected query hang.