@dantengsky 2023-08-22
> I conducted a test in the development environment and found that this PR does not have a significant impact on real use cases. I recommend merging this PR initially to determine if it resolves the issue of unexpected query hang.
> 
> cc @BohuTANG @dantengsky

confirm. 

Tested with a table consisting of 19900 blocks, point queries that hit 6 blocks show no regression.